+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Revert changes for `exercise-block` class
 * Change `Checkpoint` into `Mac Tip` in `bca`
 * Change `written-exercises` into `written-questions` in `bca`